requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
Django==4.2.9
django-ckeditor-5==0.2.18
django-phonenumber-field==8.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-dialogflow==2.41.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
idna==3.10
numpy==2.0.2
packaging==25.0
phonenumbers==9.0.9
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
